@@ -8,23 +8,24 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
-        "lunarphp/lunar": "^0.8",
-        "lunarphp/stripe": "^0.8",
+        "php": "^8.2",
         "guzzlehttp/guzzle": "^7.2",
         "http-interop/http-factory-guzzle": "^1.2",
-        "laravel/framework": "^9.52",
-        "laravel/sanctum": "^3.0",
-        "laravel/tinker": "^2.7",
-        "league/flysystem-aws-s3-v3": "^3.0"
+        "laravel/framework": "^v10.43",
+        "laravel/sanctum": "^3.3",
+        "laravel/tinker": "^2.8",
+        "league/flysystem-aws-s3-v3": "^3.0",
+        "lunarphp/lunar": "^1.0",
+        "lunarphp/table-rate-shipping": "^1.0",
+        "meilisearch/meilisearch-php": "^1.6",
+        "predis/predis": "^2.2"
     },
     "require-dev": {
         "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.15",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^6.3",
-        "phpunit/phpunit": "^9.5.10",
-        "spatie/laravel-ignition": "^1.4"
+        "phpunit/phpunit": "^10.1",
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
