@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Livewire\Components;
+namespace App\Livewire\Components;
 
+use Illuminate\View\View;
 use Livewire\Component;
 use Lunar\Base\Purchasable;
 use Lunar\Facades\CartSession;
@@ -10,42 +11,36 @@ class AddToCart extends Component
 {
     /**
      * The purchasable model we want to add to the cart.
-     *
-     * @var Purchasable
      */
     public ?Purchasable $purchasable = null;
 
     /**
      * The quantity to add to cart.
-     *
-     * @var int
      */
-    public $quantity = 1;
+    public int $quantity = 1;
 
-    /**
-     * {@inheritDoc}
-     */
-    public function rules()
+    public function rules(): array
     {
         return [
             'quantity' => 'required|numeric|min:1|max:10000',
         ];
     }
 
-    public function addToCart()
+    public function addToCart(): void
     {
         $this->validate();
 
         if ($this->purchasable->stock < $this->quantity) {
             $this->addError('quantity', 'The quantity exceeds the available stock.');
+
             return;
         }
 
         CartSession::manager()->add($this->purchasable, $this->quantity);
-        $this->emit('add-to-cart');
+        $this->dispatch('add-to-cart');
     }
 
-    public function render()
+    public function render(): View
     {
         return view('livewire.components.add-to-cart');
     }
