@@ -10,7 +10,7 @@ use Lunar\Base\Traits\LunarUser;
 
 class User extends Authenticatable
 {
-    use HasApiTokens, HasFactory, Notifiable, LunarUser;
+    use HasApiTokens, HasFactory, LunarUser, Notifiable;
 
     /**
      * The attributes that are mass assignable.
