@@ -8,8 +8,6 @@ trait FetchesUrls
 {
     /**
      * The URL model from the slug.
-     *
-     * @var \Lunar\Models\Url
      */
     public ?Url $url = null;
 
@@ -19,9 +17,8 @@ trait FetchesUrls
      * @param  string  $slug
      * @param  string  $type
      * @param  array  $eagerLoad
-     * @return \Lunar\Models\Url|null
      */
-    public function fetchUrl($slug, $type, $eagerLoad = [])
+    public function fetchUrl($slug, $type, $eagerLoad = []): ?Url
     {
         return Url::whereElementType($type)
             ->whereDefault(true)
