@@ -13,9 +13,8 @@ class CollectionSeeder extends AbstractSeeder
     /**
      * Run the database seeds.
      *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         $collections = $this->getSeedData('collections');
 
