@@ -9,15 +9,15 @@ class DatabaseSeeder extends Seeder
     /**
      * Seed the application's database.
      *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         $this->call(CollectionSeeder::class);
         $this->call(AttributeSeeder::class);
         $this->call(TaxSeeder::class);
         $this->call(ProductSeeder::class);
         $this->call(CustomerSeeder::class);
+        $this->call(ShippingSeeder::class);
         $this->call(OrderSeeder::class);
     }
 }
