@@ -22,9 +22,8 @@ class OrderSeeder extends Seeder
     /**
      * Run the database seeds.
      *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         DB::transaction(function () {
             $variants = ProductVariant::get();
@@ -71,8 +70,8 @@ class OrderSeeder extends Seeder
                                 identifier: 'VAT',
                                 description: 'VAT',
                                 percentage: 20,
-                            )
-                        ]))
+                            ),
+                        ])),
                     ]);
                 }
 
@@ -96,8 +95,8 @@ class OrderSeeder extends Seeder
                             identifier: 'VAT',
                             description: 'VAT',
                             percentage: 20,
-                        )
-                    ]))
+                        ),
+                    ])),
                 ];
 
                 if ($hasUser) {
