@@ -14,9 +14,8 @@ class TaxSeeder extends Seeder
     /**
      * Run the database seeds.
      *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         $taxClass = TaxClass::first();
 
