@@ -13,7 +13,9 @@
                 @endif
 
                 <a href="{{ route('collection.view', $this->saleCollection->defaultUrl->slug) }}"
-                   class="inline-block px-5 py-3 mt-6 text-sm font-medium text-white bg-black rounded-lg hover:ring-1 hover:ring-black">
+                   class="inline-block px-5 py-3 mt-6 text-sm font-medium text-white bg-black rounded-lg hover:ring-1 hover:ring-black"
+                   wire:navigate
+                >
                     Shop the Sale
                 </a>
             </div>
