@@ -16,10 +16,7 @@
         href="{{ asset('css/app.css') }}"
         rel="stylesheet"
     >
-    <script
-        defer
-        src="https://unpkg.com/alpinejs@3.x.x/dist/cdn.min.js"
-    ></script>
+
     <link
         rel="icon"
         href="{{ asset('favicon.svg') }}"
