@@ -1,4 +1,4 @@
-<form wire:submit.prevent="saveAddress('{{ $type }}')"
+<form wire:submit="saveAddress('{{ $type }}')"
       class="bg-white border border-gray-100 rounded-xl">
     <div class="flex items-center justify-between h-16 px-6 border-b border-gray-100">
         <h3 class="text-lg font-medium">
@@ -10,7 +10,7 @@
                 <input class="w-5 h-5 text-green-600 border-gray-100 rounded"
                        type="checkbox"
                        value="1"
-                       wire:model.defer="shippingIsBilling" />
+                       wire:model.live="shippingIsBilling" />
 
                 <span class="ml-2 text-xs font-medium">
                     Same as billing
@@ -35,7 +35,7 @@
                                    label="First name"
                                    :errors="$errors->get($type . '.first_name')"
                                    required>
-                        <x-input.text wire:model.defer="{{ $type }}.first_name"
+                        <x-input.text wire:model.live="{{ $type }}.first_name"
                                       required />
                     </x-input.group>
 
@@ -43,27 +43,27 @@
                                    label="Last name"
                                    :errors="$errors->get($type . '.last_name')"
                                    required>
-                        <x-input.text wire:model.defer="{{ $type }}.last_name"
+                        <x-input.text wire:model.live="{{ $type }}.last_name"
                                       required />
                     </x-input.group>
 
                     <x-input.group class="col-span-6"
                                    label="Company name"
                                    :errors="$errors->get($type . '.company_name')">
-                        <x-input.text wire:model.defer="{{ $type }}.company_name" />
+                        <x-input.text wire:model.live="{{ $type }}.company_name" />
                     </x-input.group>
 
                     <x-input.group class="col-span-6 sm:col-span-3"
                                    label="Contact phone"
                                    :errors="$errors->get($type . '.contact_phone')">
-                        <x-input.text wire:model.defer="{{ $type }}.contact_phone" />
+                        <x-input.text wire:model.live="{{ $type }}.contact_phone" />
                     </x-input.group>
 
                     <x-input.group class="col-span-6 sm:col-span-3"
                                    label="Contact email"
                                    :errors="$errors->get($type . '.contact_email')"
                                    required>
-                        <x-input.text wire:model.defer="{{ $type }}.contact_email"
+                        <x-input.text wire:model.live="{{ $type }}.contact_email"
                                       type="email"
                                       required />
                     </x-input.group>
@@ -76,41 +76,41 @@
                                    label="Address line 1"
                                    :errors="$errors->get($type . '.line_one')"
                                    required>
-                        <x-input.text wire:model.defer="{{ $type }}.line_one"
+                        <x-input.text wire:model.live="{{ $type }}.line_one"
                                       required />
                     </x-input.group>
 
                     <x-input.group class="col-span-3 sm:col-span-2"
                                    label="Address line 2"
                                    :errors="$errors->get($type . '.line_two')">
-                        <x-input.text wire:model.defer="{{ $type }}.line_two" />
+                        <x-input.text wire:model.live="{{ $type }}.line_two" />
                     </x-input.group>
 
                     <x-input.group class="col-span-3 sm:col-span-2"
                                    label="Address line 3"
                                    :errors="$errors->get($type . '.line_three')">
-                        <x-input.text wire:model.defer="{{ $type }}.line_three" />
+                        <x-input.text wire:model.live="{{ $type }}.line_three" />
                     </x-input.group>
 
                     <x-input.group class="col-span-3 sm:col-span-2"
                                    label="City"
                                    :errors="$errors->get($type . '.city')"
                                    required>
-                        <x-input.text wire:model.defer="{{ $type }}.city"
+                        <x-input.text wire:model.live="{{ $type }}.city"
                                       required />
                     </x-input.group>
 
                     <x-input.group class="col-span-3 sm:col-span-2"
                                    label="State / Province"
                                    :errors="$errors->get($type . '.state')">
-                        <x-input.text wire:model.defer="{{ $type }}.state" />
+                        <x-input.text wire:model.live="{{ $type }}.state" />
                     </x-input.group>
 
                     <x-input.group class="col-span-3 sm:col-span-2"
                                    label="Postcode"
                                    :errors="$errors->get($type . '.postcode')"
                                    required>
-                        <x-input.text wire:model.defer="{{ $type }}.postcode"
+                        <x-input.text wire:model.live="{{ $type }}.postcode"
                                       required />
                     </x-input.group>
 
@@ -118,7 +118,7 @@
                                    label="Country"
                                    required>
                         <select class="w-full p-3 border border-gray-200 rounded-lg sm:text-sm"
-                                wire:model.defer="{{ $type }}.country_id">
+                                wire:model.live="{{ $type }}.country_id">
                             <option value>Select a country</option>
                             @foreach ($this->countries as $country)
                                 <option value="{{ $country->id }}"
