@@ -1,4 +1,4 @@
-<form wire:submit.prevent="saveShippingOption"
+<form wire:submit="saveShippingOption"
       class="bg-white border border-gray-100 rounded-xl">
     <div class="flex items-center justify-between h-16 px-6 border-b border-gray-100">
         <h3 class="font-medium">
@@ -22,7 +22,7 @@
                         <div>
                             <input class="hidden peer"
                                    type="radio"
-                                   wire:model.defer="chosenShipping"
+                                   wire:model.live="chosenShipping"
                                    name="shippingOption"
                                    value="{{ $option->getIdentifier() }}"
                                    id="{{ $option->getIdentifier() }}" />
@@ -30,7 +30,7 @@
                             <label class="flex items-center justify-between p-4 text-sm font-medium border border-gray-100 rounded-lg shadow-sm cursor-pointer peer-checked:border-blue-500 hover:bg-gray-50 peer-checked:ring-1 peer-checked:ring-blue-500"
                                    for="{{ $option->getIdentifier() }}">
                                 <p>
-                                    {{ $option->getDescription() }}
+                                    {{ $option->getName() }}
                                 </p>
 
                                 <p>
